fix(sampling): normalize MRV2 warmup request mappings
MRV2 sampler profiling supplies int64 request mappings. Convert them to the native categorical operator's int32 metadata format and cover both upstream mapping dtypes.

+++ vllm_ascend/worker/v2/sample/gumbel.py
@@ -267,7 +267,7 @@ def gumbel_sample(
             use_fp64,
         )
 
-    expanded_idx_mapping = expanded_idx_mapping.contiguous()
+    expanded_idx_mapping = expanded_idx_mapping.to(torch.int32).contiguous()
     pos = pos.to(torch.int64).contiguous()
     if is_drafting:
         # Use the same draft/target random-stream separation as upstream.
